// client/src/autobindings/types.ts
export const AiState = __t.enum("AiState", {
  Default: __t.unit(),
  BossEnderIdle: __t.unit(),
  BossEnderChase: __t.unit(),
  BossEnderDance: __t.unit(),
  BossEnderVanish: __t.unit(),
  BossEnderLurk: __t.unit(),
  BossEnderTeleport: __t.unit(),
  BossEnderTransform: __t.unit(),
  Stationary: __t.unit(),
  BossAgnaIdle: __t.unit(),
  BossAgnaFlamethrower: __t.unit(),
  BossAgnaMagicCircle: __t.unit(),
  BossAgnaRitualMatch: __t.unit(),
  BossAgnaRitualWick: __t.unit(),
  BossAgnaRitualFailed: __t.unit(),
  BossAgnaRitualComplete: __t.unit(),
  BossSimonIdle: __t.unit(),
  BossSimonToxicSpray: __t.unit(),
  BossSimonZombieWave: __t.unit(),
  BossSimonTransform: __t.unit(),
  BossSimonChemicalBoltPattern: __t.unit(),
  BossSimonToxicZonePattern: __t.unit(),
  BossSimonPhase2Transform: __t.unit(),
});
export type AiState = __Infer<typeof AiState>;
// ...
export const DbVector2 = __t.object("DbVector2", {
  x: __t.f32(),
  y: __t.f32(),
});
export type DbVector2 = __Infer<typeof DbVector2>;
// ...
export const Entity = __t.object("Entity", {
  entityId: __t.u32(),
  get position() {
    return DbVector2;
  },
  get direction() {
    return DbVector2;
  },
  radius: __t.f32(),
  get waypoint() {
    return DbVector2;
  },
  hasWaypoint: __t.bool(),
});
export type Entity = __Infer<typeof Entity>;
// ...
export const MonsterType = __t.enum("MonsterType", {
  Rat: __t.unit(),
  Slime: __t.unit(),
  Bat: __t.unit(),
  Orc: __t.unit(),
  Imp: __t.unit(),
  Zombie: __t.unit(),
  VoidChest: __t.unit(),
  EnderClaw: __t.unit(),
  BossEnderPhase1: __t.unit(),
  BossEnderPhase2: __t.unit(),
  BossAgnaPhase1: __t.unit(),
  BossAgnaPhase2: __t.unit(),
  BossSimonPhase1: __t.unit(),
  BossSimonPhase2: __t.unit(),
  AgnaCandle: __t.unit(),
  Crate: __t.unit(),
  Tree: __t.unit(),
  Statue: __t.unit(),
});
export type MonsterType = __Infer<typeof MonsterType>;

// The tagged union or sum type for the algebraic type `MonsterVariant`.
export const MonsterVariant = __t.enum("MonsterVariant", {
  Default: __t.unit(),
  Shiny: __t.unit(),
  Cursed: __t.unit(),
});
export type MonsterVariant = __Infer<typeof MonsterVariant>;

export const Monsters = __t.object("Monsters", {
  monsterId: __t.u32(),
  get bestiaryId() {
    return MonsterType;
  },
  get variant() {
    return MonsterVariant;
  },
  hp: __t.u32(),
  maxHp: __t.u32(),
  atk: __t.f32(),
  speed: __t.f32(),
  targetPlayerId: __t.u32(),
  get aiState() {
    return AiState;
  },
  radius: __t.f32(),
  get spawnPosition() {
    return DbVector2;
  },
});
export type Monsters = __Infer<typeof Monsters>;